separated AppWindow and CloseButton
