+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/Users/aberllin/Desktop/react-project/src/reportWebVitals.js":"1","/Users/aberllin/Desktop/react-project/src/App.js":"2","/Users/aberllin/Desktop/react-project/src/modules/Dashboard.js":"3","/Users/aberllin/Desktop/react-project/src/components/AppCard.js":"4","/Users/aberllin/Desktop/react-project/src/modules/appsData.js":"5","/Users/aberllin/Desktop/react-project/src/index.js":"6","/Users/aberllin/Desktop/react-project/src/components/Toggle.js":"7","/Users/aberllin/Desktop/react-project/src/styles/theme.js":"8","/Users/aberllin/Desktop/react-project/src/styles/global.js":"9"},{"size":362,"mtime":1608128016636,"results":"10","hashOfConfig":"11"},{"size":720,"mtime":1609702117871,"results":"12","hashOfConfig":"11"},{"size":1655,"mtime":1609762194877,"results":"13","hashOfConfig":"11"},{"size":2050,"mtime":1609762127647,"results":"14","hashOfConfig":"11"},{"size":906,"mtime":1609762485644,"results":"15","hashOfConfig":"11"},{"size":442,"mtime":1608918543698,"results":"16","hashOfConfig":"11"},{"size":891,"mtime":1609701971354,"results":"17","hashOfConfig":"11"},{"size":347,"mtime":1609701971359,"results":"18","hashOfConfig":"11"},{"size":302,"mtime":1608978951785,"results":"19","hashOfConfig":"11"},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},"1vzt8v8",{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"31","messages":"32","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"33"},{"filePath":"34","messages":"35","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"36","messages":"37","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"38","messages":"39","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/aberllin/Desktop/react-project/src/reportWebVitals.js",[],["40","41"],"/Users/aberllin/Desktop/react-project/src/App.js",[],"/Users/aberllin/Desktop/react-project/src/modules/Dashboard.js",[],"/Users/aberllin/Desktop/react-project/src/components/AppCard.js",[],"/Users/aberllin/Desktop/react-project/src/modules/appsData.js",[],"/Users/aberllin/Desktop/react-project/src/index.js",[],["42","43"],"/Users/aberllin/Desktop/react-project/src/components/Toggle.js",[],"/Users/aberllin/Desktop/react-project/src/styles/theme.js",[],"/Users/aberllin/Desktop/react-project/src/styles/global.js",[],{"ruleId":"44","replacedBy":"45"},{"ruleId":"46","replacedBy":"47"},{"ruleId":"44","replacedBy":"48"},{"ruleId":"46","replacedBy":"49"},"no-native-reassign",["50"],"no-negated-in-lhs",["51"],["50"],["51"],"no-global-assign","no-unsafe-negation"]
+[{"/Users/aberllin/Desktop/react-project/src/reportWebVitals.js":"1","/Users/aberllin/Desktop/react-project/src/App.js":"2","/Users/aberllin/Desktop/react-project/src/modules/Dashboard.js":"3","/Users/aberllin/Desktop/react-project/src/components/AppCard.js":"4","/Users/aberllin/Desktop/react-project/src/modules/appsData.js":"5","/Users/aberllin/Desktop/react-project/src/index.js":"6","/Users/aberllin/Desktop/react-project/src/components/Toggle.js":"7","/Users/aberllin/Desktop/react-project/src/styles/theme.js":"8","/Users/aberllin/Desktop/react-project/src/styles/global.js":"9","/Users/aberllin/Desktop/react-project/src/components/AppWindow.js":"10","/Users/aberllin/Desktop/react-project/src/components/CloseAppButton.js":"11"},{"size":362,"mtime":1608128016636,"results":"12","hashOfConfig":"13"},{"size":720,"mtime":1609702117871,"results":"14","hashOfConfig":"13"},{"size":1655,"mtime":1609762194877,"results":"15","hashOfConfig":"13"},{"size":1264,"mtime":1609869277601,"results":"16","hashOfConfig":"13"},{"size":906,"mtime":1609867304288,"results":"17","hashOfConfig":"13"},{"size":442,"mtime":1609696509826,"results":"18","hashOfConfig":"13"},{"size":891,"mtime":1609701971354,"results":"19","hashOfConfig":"13"},{"size":347,"mtime":1609701971359,"results":"20","hashOfConfig":"13"},{"size":302,"mtime":1609701348546,"results":"21","hashOfConfig":"13"},{"size":655,"mtime":1609867761138,"results":"22","hashOfConfig":"13"},{"size":419,"mtime":1609867717897,"results":"23","hashOfConfig":"13"},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},"1vzt8v8",{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"31","messages":"32","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"35","messages":"36","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"37","messages":"38","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"39","messages":"40","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"41","messages":"42","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"43","messages":"44","errorCount":1,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"45","messages":"46","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/aberllin/Desktop/react-project/src/reportWebVitals.js",[],["47","48"],"/Users/aberllin/Desktop/react-project/src/App.js",[],"/Users/aberllin/Desktop/react-project/src/modules/Dashboard.js",[],"/Users/aberllin/Desktop/react-project/src/components/AppCard.js",[],"/Users/aberllin/Desktop/react-project/src/modules/appsData.js",[],"/Users/aberllin/Desktop/react-project/src/index.js",[],"/Users/aberllin/Desktop/react-project/src/components/Toggle.js",[],"/Users/aberllin/Desktop/react-project/src/styles/theme.js",[],"/Users/aberllin/Desktop/react-project/src/styles/global.js",[],"/Users/aberllin/Desktop/react-project/src/components/AppWindow.js",["49"],"/Users/aberllin/Desktop/react-project/src/components/CloseAppButton.js",[],{"ruleId":"50","replacedBy":"51"},{"ruleId":"52","replacedBy":"53"},{"ruleId":"54","severity":2,"message":"55","line":9,"column":10,"nodeType":"56","endLine":9,"endColumn":24},"no-native-reassign",["57"],"no-negated-in-lhs",["58"],"react/jsx-no-undef","'CloseAppButton' is not defined.","JSXIdentifier","no-global-assign","no-unsafe-negation"]
