+++ src/components/AppCard.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import styled from 'styled-components'
-import { Link, Route } from 'react-router-dom'
-import { IoCloseCircleOutline } from 'react-icons/io5'
+import { Route, Link } from 'react-router-dom'
+import { AppWindow } from './AppWindow'
 
 export const AppCard = ({ image, title, app }) => {
   const urlTitle = title
@@ -26,48 +26,6 @@ export const AppCard = ({ image, title, app }) => {
 const StyledLink = styled(Link)`
   text-decoration: none;
   color: ${({ theme }) => theme.textColor};
-`
-
-const AppWindow = ({ title, app }) => {
-  return (
-    <AppWindowWrapper>
-      <AppCardNav>
-        <div>{title}</div>
-        <Link to="/">
-          <CloseIcon />
-        </Link>
-      </AppCardNav>
-      <div>{app}</div>
-    </AppWindowWrapper>
-  )
-}
-
-const CloseIcon = styled(IoCloseCircleOutline)`
-  color: ${({ theme }) => theme.textColor};
-  transition: ease 0.2s;
-
-  &:hover {
-    color: red;
-  }
-`
-
-const AppCardNav = styled.div`
-  display: flex;
-  justify-content: space-around;
-  padding: 50px;
-`
-
-const AppWindowWrapper = styled.div`
-  position: fixed;
-  z-index: 2;
-  left: 0;
-  top: 0;
-  width: 100%;
-  height: 100%;
-  overflow: auto;
-  font-size: 35px;
-  color: ${({ theme }) => theme.textColor};
-  background-color: ${({ theme }) => theme.bodyColor};
 `
 
 const Image = styled.div`
