+++ src/components/CloseAppButton.js
@@ -1,0 +1,21 @@
+import React from 'react'
+import styled from 'styled-components'
+import { Link } from 'react-router-dom'
+import { IoCloseCircleOutline } from 'react-icons/io5'
+
+export const CloseAppButton = () => {
+  return (
+    <Link to="/">
+      <CloseIcon />
+    </Link>
+  )
+}
+
+const CloseIcon = styled(IoCloseCircleOutline)`
+  color: ${({ theme }) => theme.textColor};
+  transition: ease 0.2s;
+
+  &:hover {
+    color: red;
+  }
+`
